+++ megfile/s3_path.py
@@ -7,16 +7,18 @@ from concurrent.futures import ThreadPoolExecutor
 from functools import cached_property, lru_cache, wraps
 from logging import getLogger as get_logger
 from typing import IO, Any, BinaryIO, Callable, Dict, Iterator, List, Optional, Tuple
+from urllib.parse import urlparse
 
 import boto3
 import botocore
-from botocore.awsrequest import AWSResponse
+from botocore.awsrequest import AWSPreparedRequest, AWSResponse
 
 from megfile.config import (
     DEFAULT_BLOCK_SIZE,
     DEFAULT_MAX_BLOCK_SIZE,
     DEFAULT_MIN_BLOCK_SIZE,
     GLOBAL_MAX_WORKERS,
+    HTTP_AUTH_HEADERS,
     S3_CLIENT_CACHE_MODE,
     S3_MAX_RETRY_TIMES,
 )
@@ -76,6 +78,7 @@ from megfile.utils import (
     generate_cache_path,
     get_binary_mode,
     get_content_offset,
+    is_domain_or_subdomain,
     is_readable,
     necessary_params,
     process_local,
@@ -110,7 +113,7 @@ max_retries = S3_MAX_RETRY_TIMES
 max_keys = 1000
 
 
-def _patch_make_request(client: botocore.client.BaseClient):
+def _patch_make_request(client: botocore.client.BaseClient, redirect: bool = False):
     def after_callback(result: Tuple[AWSResponse, dict], *args, **kwargs):
         if (
             not isinstance(result, tuple)
@@ -149,6 +152,32 @@ def _patch_make_request(client: botocore.client.BaseClient):
         before_callback=before_callback,
         retry_callback=retry_callback,
     )
+
+    def patch_send(send):
+        def patched_send(request: AWSPreparedRequest) -> AWSResponse:
+            response: AWSResponse = send(request)
+            if (
+                request.method == "GET"  # only support GET method for now
+                and response.status_code in (301, 302, 307, 308)
+                and "Location" in response.headers
+            ):
+                # Permit sending auth/cookie headers from "foo.com" to "sub.foo.com".
+                # See also: https://go.dev/src/net/http/client.go#L980
+                location = response.headers["Location"]
+                ihost = urlparse(request.url).hostname
+                dhost = urlparse(location).hostname
+                if not is_domain_or_subdomain(dhost, ihost):
+                    for name in HTTP_AUTH_HEADERS:
+                        request.headers.pop(name, None)
+                request.url = location
+                response = send(request)
+            return response
+
+        return patched_send
+
+    if redirect:
+        client._endpoint._send = patch_send(client._endpoint._send)
+
     return client
 
 
@@ -168,7 +197,11 @@ def parse_s3_url(s3_url: PathLike) -> Tuple[str, str]:
 
 
 def get_scoped_config(profile_name: Optional[str] = None) -> Dict:
-    return get_s3_session(profile_name=profile_name)._session.get_scoped_config()
+    try:
+        session = get_s3_session(profile_name=profile_name)
+    except botocore.exceptions.ProfileNotFound:
+        session = get_s3_session()
+    return session._session.get_scoped_config()
 
 
 @lru_cache()
@@ -212,25 +245,22 @@ def get_s3_session(profile_name=None) -> boto3.Session:
     )
 
 
+def get_env_var(env_name: str, profile_name=None):
+    if profile_name:
+        return os.getenv(f"{profile_name}__{env_name}".upper())
+    return os.getenv(env_name.upper())
+
+
+def parse_boolean(value: Optional[str], default: bool = False) -> bool:
+    if value is None:
+        return default
+    return value.lower() in ("true", "yes", "1")
+
+
 def get_access_token(profile_name=None):
-    access_key_env_name = (
-        f"{profile_name}__AWS_ACCESS_KEY_ID".upper()
-        if profile_name
-        else "AWS_ACCESS_KEY_ID"
-    )
-    secret_key_env_name = (
-        f"{profile_name}__AWS_SECRET_ACCESS_KEY".upper()
-        if profile_name
-        else "AWS_SECRET_ACCESS_KEY"
-    )
-    session_token_env_name = (
-        f"{profile_name}__AWS_SESSION_TOKEN".upper()
-        if profile_name
-        else "AWS_SESSION_TOKEN"
-    )
-    access_key = os.getenv(access_key_env_name)
-    secret_key = os.getenv(secret_key_env_name)
-    session_token = os.getenv(session_token_env_name)
+    access_key = get_env_var("AWS_ACCESS_KEY_ID", profile_name=profile_name)
+    secret_key = get_env_var("AWS_SECRET_ACCESS_KEY", profile_name=profile_name)
+    session_token = get_env_var("AWS_SESSION_TOKEN", profile_name=profile_name)
     if access_key and secret_key:
         return access_key, secret_key, session_token
 
@@ -277,10 +307,7 @@ def get_s3_client(
             connect_timeout=5, max_pool_connections=GLOBAL_MAX_WORKERS
         )
 
-    addressing_style_env_key = "AWS_S3_ADDRESSING_STYLE"
-    if profile_name:
-        addressing_style_env_key = f"{profile_name}__AWS_S3_ADDRESSING_STYLE".upper()
-    addressing_style = os.environ.get(addressing_style_env_key)
+    addressing_style = get_env_var("AWS_S3_ADDRESSING_STYLE", profile_name=profile_name)
     if addressing_style:
         config = config.merge(
             botocore.config.Config(s3={"addressing_style": addressing_style})
@@ -291,15 +318,25 @@ def get_s3_client(
         session = get_s3_session(profile_name=profile_name)
     except botocore.exceptions.ProfileNotFound:
         session = get_s3_session()
+
+    s3_config = get_scoped_config(profile_name=profile_name).get("s3", {})
+    verify = get_env_var("AWS_S3_VERIFY", profile_name=profile_name)
+    verify = verify or s3_config.get("verify")
+    verify = parse_boolean(verify, default=True)
+    redirect = get_env_var("AWS_S3_REDIRECT", profile_name=profile_name)
+    redirect = redirect or s3_config.get("redirect")
+    redirect = parse_boolean(redirect, default=False)
+
     client = session.client(
         "s3",
         endpoint_url=get_endpoint_url(profile_name=profile_name),
+        verify=verify,
         config=config,
         aws_access_key_id=access_key,
         aws_secret_access_key=secret_key,
         aws_session_token=session_token,
     )
-    client = _patch_make_request(client)
+    client = _patch_make_request(client, redirect=redirect)
     return client